Volume Changes - New volume operation mode

+++ app/src/main/java/fgapps/com/br/iassistant2/interfaces/VolumeChangeListener.kt
@@ -2,6 +2,6 @@ package fgapps.com.br.iassistant2.interfaces
 
 interface VolumeChangeListener {
 
-    fun volumeChange(volume: Int)
+    fun volumeChange(volume: Float)
 
 }
